リファクタリング:終了.
